first implementation
In this commit we implement the simplest version of custom malloc called , mymalloc and myfree
we define a struct called BlockHeader that we use to store the size of the memory block allocated.
We will need in myfree.

In the main section we calculate the performance of the default malloc against the custom one.
This will be later moved to a benchmark file we will generate.